+++ packages/client/src/middlewares/actionDelegation.ts
@@ -1,15 +1,15 @@
 import { statesHookHelper } from '@/util/helper';
 import {
+  $self,
   createAssert,
-  falseValue,
   filterItem,
   forEach,
   instanceOf,
   isNumber,
   isString,
   objectKeys,
-  pushItem,
-  trueValue
+  objectValues,
+  pushItem
 } from '@alova/shared';
 import { AlovaGenerics, promiseStatesHook } from 'alova';
 import {
@@ -19,7 +19,9 @@ import {
   AlovaGuardNext
 } from '~/typings/clienthook';
 
-const actionsMap: Record<string | number | symbol, Actions[]> = {};
+let currentHookIndex = 0;
+// (id, (hookIndex, Actions))
+const actionsMap: Record<string | number | symbol, Record<string | number, Actions>> = {};
 const isFrontMiddlewareContext = <AG extends AlovaGenerics = AlovaGenerics, Args extends any[] = any[]>(
   context: AlovaFrontMiddlewareContext<AG, Args> | AlovaFetcherMiddlewareContext<AG, Args>
 ): context is AlovaFrontMiddlewareContext<AG, Args> => !!(context as AlovaFrontMiddlewareContext<AG, Args>).send;
@@ -37,9 +39,21 @@ const assert = createAssert('subscriber');
 export const actionDelegationMiddleware = <AG extends AlovaGenerics = AlovaGenerics, Args extends any[] = any[]>(
   id: string | number | symbol
 ) => {
-  const { ref } = statesHookHelper(promiseStatesHook());
+  const { ref, onUnmounted } = statesHookHelper(promiseStatesHook());
 
-  const delegated = ref(falseValue);
+  const hookIndex = ref(currentHookIndex + 1);
+
+  if (hookIndex.current > currentHookIndex) {
+    currentHookIndex += 1;
+  }
+
+  onUnmounted(() => {
+    if (actionsMap[id]?.[hookIndex.current]) {
+      // delete action on unmount
+      delete actionsMap[id][hookIndex.current];
+    }
+  });
+
   return (
     context: (AlovaFrontMiddlewareContext<AG, Args> | AlovaFetcherMiddlewareContext<AG, Args>) & {
       delegatingActions?: Actions;
@@ -47,34 +61,31 @@ export const actionDelegationMiddleware = <AG extends AlovaGenerics = AlovaGener
     next: AlovaGuardNext<AG, Args>
   ) => {
     // The middleware will be called repeatedly. If you have already subscribed, you do not need to subscribe again.
-    if (!delegated.current) {
-      const { abort, proxyStates, delegatingActions = {} } = context;
-      const update = (newStates: Record<string, any>) => {
-        type ProxyStateKeys = keyof typeof proxyStates;
-        for (const key in newStates) {
-          proxyStates[key as ProxyStateKeys] && (proxyStates[key as ProxyStateKeys].v = newStates[key]);
+    const { abort, proxyStates, delegatingActions = {} } = context;
+    const update = (newStates: Record<string, any>) => {
+      type ProxyStateKeys = keyof typeof proxyStates;
+      for (const key in newStates) {
+        proxyStates[key as ProxyStateKeys] && (proxyStates[key as ProxyStateKeys].v = newStates[key]);
+      }
+    };
+    // Those with the same ID will be saved together in the form of an array
+    const hooks = (actionsMap[id] = actionsMap[id] || []);
+    const handler = isFrontMiddlewareContext(context)
+      ? {
+          ...delegatingActions,
+          send: context.send,
+          abort,
+          update
         }
-      };
-      // Those with the same ID will be saved together in the form of an array
-      const handlersItems = (actionsMap[id] = actionsMap[id] || []);
-      handlersItems.push(
-        isFrontMiddlewareContext(context)
-          ? {
-              ...delegatingActions,
-              send: context.send,
-              abort,
-              update
-            }
-          : {
-              ...delegatingActions,
-              fetch: context.fetch,
-              abort,
-              update
-            }
-      );
+      : {
+          ...delegatingActions,
+          fetch: context.fetch,
+          abort,
+          update
+        };
 
-      delegated.current = trueValue;
-    }
+    hooks[hookIndex.current] = handler;
+
     return next();
   };
 };
@@ -92,12 +103,12 @@ export const accessAction = (
 ) => {
   const matched = [] as Actions[];
   if (typeof id === 'symbol' || isString(id) || isNumber(id)) {
-    actionsMap[id] && pushItem(matched, ...actionsMap[id]);
+    actionsMap[id] && pushItem(matched, ...objectValues(actionsMap[id]));
   } else if (instanceOf(id, RegExp)) {
     forEach(
       filterItem(objectKeys(actionsMap), idItem => id.test(idItem)),
       idItem => {
-        pushItem(matched, ...actionsMap[idItem]);
+        pushItem(matched, ...objectValues(actionsMap[idItem]));
       }
     );
   }
@@ -107,5 +118,5 @@ export const accessAction = (
     assert(false, `no handler can be matched by using \`${id.toString()}\``);
   }
 
-  forEach(matched, onMatch);
+  forEach(filterItem(matched, $self), onMatch);
 };