+++ packages/vue-options/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @alova/vue-options
 
+## 2.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`9e95069`](https://github.com/alovajs/alova/commit/9e95069b0fee0e86edb750f4bd5ab2c60083bcaa)]:
+  - @alova/shared@1.1.2
+
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`cf0b04a`](https://github.com/alovajs/alova/commit/cf0b04a82829c116232d487c046fec3c563f8a8e)]:
+  - @alova/shared@1.1.1
+
 ## 2.0.8
 
 ### Patch Changes